requirements.txt
django==4.1.13
djongo==1.3.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
requests==2.31.0
mongoengine==0.27.0
pymongo==3.12.1
dnspython==2.6.1